+++ src/game/assets.ts
@@ -13,8 +13,8 @@ import {
   unshareProvokingVertices,
 } from "../render/mesh.js";
 import { AABB } from "../physics/broadphase.js";
-import { RendererDef } from "../render/render-init.js";
-import { Renderer } from "../render/renderer.js";
+import { RendererDef } from "../render/renderer-ecs.js";
+import { Renderer } from "../render/renderer-ecs.js";
 import { assert } from "../test.js";
 import { objMap } from "../util.js";
 import { getText } from "../webget.js";
