+++ src/game/boat.ts
@@ -3,7 +3,7 @@ import { PhysicsTimerDef } from "../time.js";
 import { quat, vec3 } from "../gl-matrix.js";
 import { jitter } from "../math.js";
 import { ColorDef } from "../color.js";
-import { RenderableConstructDef } from "../render/renderer.js";
+import { RenderableConstructDef } from "../render/renderer-ecs.js";
 import {
   PhysicsParentDef,
   Position,
