+++ src/game/enemy.ts
@@ -14,7 +14,7 @@ import {
   ScaleDef,
 } from "../physics/transform.js";
 import { scaleMesh3 } from "../render/mesh.js";
-import { RenderableConstructDef } from "../render/renderer.js";
+import { RenderableConstructDef } from "../render/renderer-ecs.js";
 import { Assets } from "./assets.js";
 import { ColorDef } from "../color.js";
