+++ src/game/grapple.ts
@@ -10,7 +10,7 @@ import {
   PositionDef,
   RotationDef,
 } from "../physics/transform.js";
-import { RenderableConstructDef } from "../render/renderer.js";
+import { RenderableConstructDef } from "../render/renderer-ecs.js";
 import { AssetsDef } from "./assets.js";
 import { ColorDef } from "../color.js";
 import { InteractableDef } from "./interact.js";
