+++ src/game/player.ts
@@ -6,7 +6,10 @@ import { EM, Entity, EntityManager, EntityW } from "../entity-manager.js";
 import { PhysicsTimerDef } from "../time.js";
 import { ColorDef } from "../color.js";
 import { FinishedDef } from "../build.js";
-import { RenderableConstructDef, RenderableDef } from "../render/renderer.js";
+import {
+  RenderableConstructDef,
+  RenderableDef,
+} from "../render/renderer-ecs.js";
 import {
   PhysicsParentDef,
   PositionDef,
