+++ src/game/sandbox.ts
@@ -24,8 +24,11 @@ import {
   ScaleDef,
 } from "../physics/transform.js";
 import { cloneMesh, scaleMesh } from "../render/mesh.js";
-import { RenderableDef, RenderableConstructDef } from "../render/renderer.js";
-import { RendererDef } from "../render/render-init.js";
+import {
+  RenderableDef,
+  RenderableConstructDef,
+} from "../render/renderer-ecs.js";
+import { RendererDef } from "../render/renderer-ecs.js";
 import { tempVec } from "../temp-pool.js";
 import { assert } from "../test.js";
 import { TimeDef } from "../time.js";
